+++ VasysRomanNumeralsKata/RomanNumeral.cs
@@ -7,17 +7,17 @@ namespace VasysRomanNumeralsKata
 {
     public class RomanNumeral
     {
-        private int? _baseTenRepresentation = null;
-        private int remainder = 0;
+        private long? _baseTenRepresentation = null;
+        private long remainder = 0;
         private StringBuilder romanNumeralResult = new StringBuilder();
 
-        public RomanNumeral(int baseTenNumber)
+        public RomanNumeral(long baseTenNumber)
         {
             _baseTenRepresentation = baseTenNumber;
-            int largestSupportedValue = (LargestNumeral() * 4) - 1;
+            long largestSupportedValue = (LargestNumeral() * 4) - 1;
             if (_baseTenRepresentation < 1 || _baseTenRepresentation > largestSupportedValue)
                 throw new ArgumentOutOfRangeException("Value cannot be represented with the given set of roman numeral values");
-            remainder = (int)_baseTenRepresentation;
+            remainder = (long)_baseTenRepresentation;
         }
 
         // Values can be added, provided that they are added in pairs, where the first value is 
@@ -25,7 +25,7 @@ namespace VasysRomanNumeralsKata
         //  and the second value is the first value divided by the 
         //  factorDifferenceBetweenRepeatableNumeralAndPartialStep variable (originally 2)
         //  (e.g. if 1000 is the largest existing value, then 10000 and 5000 can be added).
-        private static readonly Dictionary<int, char> numeralLookup = new Dictionary<int, char>()
+        private static readonly Dictionary<long, char> numeralLookup = new Dictionary<long, char>()
         {
             {1000, 'M'},
             {500, 'D'},
@@ -39,17 +39,17 @@ namespace VasysRomanNumeralsKata
         // For theoretical future use, these values could be configurable.  Radix is much more likely to be configured,
         //  but factorDifferenceBetweenRepeatableNumeralAndPartialStep might be considered for larger radix values.
         //  The radix must always be evenly divisible by the factorDifferenceBetweenRepeatableNumeralAndPartialStep.
-        private static readonly int radix = 10;
-        private static readonly int factorDifferenceBetweenRepeatableNumeralAndPartialStep = 2;
+        private static readonly long radix = 10;
+        private static readonly long factorDifferenceBetweenRepeatableNumeralAndPartialStep = 2;
 
-        private static int LargestNumeral()
+        private static long LargestNumeral()
         {
             return numeralLookup.Keys.Max();
         }
 
         public string GenerateRomanNumeralRepresentation()
         {
-            int currentNumeral = LargestNumeral();
+            long currentNumeral = LargestNumeral();
 
             // I originally considered making this a recursive call, but determined that would add
             //  unnecessary complexity.  
@@ -67,16 +67,16 @@ namespace VasysRomanNumeralsKata
             return romanNumeralResult.ToString();
         }
 
-        private void CalculateNumeralsForRepeatableNumeralValue(int repeatableValue)
+        private void CalculateNumeralsForRepeatableNumeralValue(long repeatableValue)
         {
-            int decrementNumeralValue = repeatableValue / radix;
+            long decrementNumeralValue = repeatableValue / radix;
             CalculateNumeralsForNumeralValue(repeatableValue, decrementNumeralValue);
 
-            int partialStepNumeralValue = repeatableValue / factorDifferenceBetweenRepeatableNumeralAndPartialStep;
+            long partialStepNumeralValue = repeatableValue / factorDifferenceBetweenRepeatableNumeralAndPartialStep;
             CalculateNumeralsForNumeralValue(partialStepNumeralValue, decrementNumeralValue);
         }
 
-        private void CalculateNumeralsForNumeralValue(int numeralValue, int decrementNumeralValue)
+        private void CalculateNumeralsForNumeralValue(long numeralValue, long decrementNumeralValue)
         {
             char numeralCharacter = numeralLookup[numeralValue];
             while (remainder >= numeralValue)
@@ -86,7 +86,7 @@ namespace VasysRomanNumeralsKata
             }
 
             char decrementNumeralCharacter = numeralLookup[decrementNumeralValue];
-            int numeralValueMinusDecrement = numeralValue - decrementNumeralValue;
+            long numeralValueMinusDecrement = numeralValue - decrementNumeralValue;
             if (remainder >= numeralValueMinusDecrement)
             {
                 romanNumeralResult.Append(decrementNumeralCharacter);
